Fix multiple token search issue
In Go, the loop iterator variable is a single variable that takes different values in each loop iteration.
When we try to take a reference of the iteration var, its value changes with each iteration. This results in the loop always "finding" the final element.
To fix this, we can make a copy of the `searchToken` variable and take its reference instead.

https://github.com/golang/go/wiki/CommonMistakes#using-reference-to-loop-iterator-variable