Fix doppler_secrets data source to use ID based on project+config
When the `doppler_secrets` resource was written originally, it did not
have options for specifying a project or config because it was intended
to be used with service tokens. An ID must be set for Terraform data
sources so we just used the API client's ID.

Once we added project+config as fields, it was possible to create
multiple data sources with the same ID. This does not appear to cause
problems in fetching data but is still incorrect.

The fix here uses the project+config as the ID for the resource,
ensuring uniqueness.

This change will not break users with existing `doppler_secrets` data
sources in their state. They likely won't even notice the change in
their Terraform plan/apply unless they are referencing the data source
ID somewhere (which is unlikely because it is not useful).